ISSUE #3021 - GroupList component Splitted into different component files
